refactor : rebook-product 모듈 생
