+++ rebook-product/src/main/java/com/be/rebook/product/dto/ProductRequestDTO.java
@@ -1,4 +1,4 @@
-package com.be.rebook.domain.product.dto;
+package com.be.rebook.product.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
