+++ rebook-product/src/main/java/com/be/rebook/product/entity/Product.java
@@ -1,8 +1,8 @@
-package com.be.rebook.domain.product.entity;
+package com.be.rebook.product.entity;
 
 import com.be.rebook.domain.members.entity.Members;
-import com.be.rebook.domain.product.dto.ProductRequestDTO;
-import com.be.rebook.domain.product.type.ProductStatus;
+import com.be.rebook.product.dto.ProductRequestDTO;
+import com.be.rebook.product.type.ProductStatus;
 import com.be.rebook.global.config.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
@@ -53,7 +53,7 @@ public class Product extends BaseEntity {
     private List<ProductImage> productImages = new ArrayList<>();
 
     public static Product of(Members member, ProductRequestDTO.ProductSaveRequestDTO productSaveRequestDTO) {
-        return Product.builder()
+        return builder()
                 .title(productSaveRequestDTO.getTitle())
                 .content(productSaveRequestDTO.getContent())
                 .price(productSaveRequestDTO.getPrice())
