+++ rebook-product/src/main/java/com/be/rebook/product/entity/ProductImage.java
@@ -1,6 +1,5 @@
-package com.be.rebook.domain.product.entity;
+package com.be.rebook.product.entity;
 
-import com.be.rebook.domain.product.entity.Product;
 import com.be.rebook.global.config.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
