+++ rebook-product/src/main/java/com/be/rebook/product/repository/ProductImageRepository.java
@@ -1,6 +1,6 @@
-package com.be.rebook.domain.product.repository;
+package com.be.rebook.product.repository;
 
-import com.be.rebook.domain.product.entity.ProductImage;
+import com.be.rebook.product.entity.ProductImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
