+++ rebook-product/src/main/java/com/be/rebook/product/repository/ProductRepository.java
@@ -1,7 +1,7 @@
-package com.be.rebook.domain.product.repository;
+package com.be.rebook.product.repository;
 
 import com.be.rebook.domain.members.entity.Members;
-import com.be.rebook.domain.product.entity.Product;
+import com.be.rebook.product.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
