+++ rebook-product/src/main/java/com/be/rebook/product/service/ProductService.java
@@ -1,14 +1,14 @@
-package com.be.rebook.domain.product.service;
+package com.be.rebook.product.service;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.be.rebook.domain.members.entity.Members;
 import com.be.rebook.domain.members.repository.MembersRepository;
-import com.be.rebook.domain.product.entity.Product;
-import com.be.rebook.domain.product.dto.ProductRequestDTO;
-import com.be.rebook.domain.product.repository.ProductRepository;
-import com.be.rebook.domain.product.entity.ProductImage;
-import com.be.rebook.domain.product.repository.ProductImageRepository;
+import com.be.rebook.product.entity.Product;
+import com.be.rebook.product.dto.ProductRequestDTO;
+import com.be.rebook.product.repository.ProductRepository;
+import com.be.rebook.product.entity.ProductImage;
+import com.be.rebook.product.repository.ProductImageRepository;
 import com.be.rebook.global.exception.BaseException;
 import com.be.rebook.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
@@ -23,7 +23,7 @@ import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
-import static com.be.rebook.domain.product.dto.ProductResponseDTO.*;
+import static com.be.rebook.product.dto.ProductResponseDTO.*;
 
 @Service
 @RequiredArgsConstructor
