+++ rebook-product/src/main/java/com/be/rebook/product/type/ProductStatus.java
@@ -1,4 +1,4 @@
-package com.be.rebook.domain.product.type;
+package com.be.rebook.product.type;
 
 public enum ProductStatus {
     PENDING, COMPLETED
