+++ settings.gradle
@@ -2,3 +2,5 @@ rootProject.name = 'reBook'
 include 'rebook-members'
 include 'rebook-auth'
 include 'rebook-common'
+include 'rebook-product'
+
